requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
branca==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
crispy-bootstrap4==2023.1
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
fiona==1.9.5
flask==3.0.0
folium==0.15.0
geopandas==0.13.2
idna==3.4
importlib-metadata==6.8.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg==3.1.13
psycopg2==2.9.9
pure-eval==0.2.2
pygments==2.17.1
PyJWT==2.8.0
pykml==0.2.0
pyproj
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
validate-email==1.3
wcwidth==0.2.11
werkzeug==3.0.1
zipp==3.17.0
